stm32: flash: pull out i/d cache support.
Copied a few times.
